tests/wpt/tests/css/css-position/parsing/z-index-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Positioned Layout Module Level 3: parsing z-index with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-position/#propdef-z-index">
<meta name="assert" content="z-index supports only the grammar 'auto | <integer>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("z-index", "none");
test_invalid_value("z-index", "10px");
test_invalid_value("z-index", "0.5");
test_invalid_value("z-index", "auto 123");
</script>
</body>
</html>
